update: screens folder structure
