+++ frontend/student_apk/lib/main.dart
@@ -3,21 +3,21 @@ import 'package:student_apk/routes.dart';
 import 'package:student_apk/screens/loading.dart';
 import 'package:student_apk/screens/get_started_page.dart';
 import 'package:student_apk/screens/login.dart';
-import 'package:student_apk/screens/signup_step1_page.dart';
-import 'package:student_apk/screens/signup_step2_page.dart';
-import 'package:student_apk/screens/signup_step3_page.dart';
-import 'package:student_apk/screens/signup_step4_page.dart';
-import 'package:student_apk/screens/verification_complete_page.dart';
+import 'package:student_apk/screens/sign_up_pages/signup_step1_page.dart';
+import 'package:student_apk/screens/sign_up_pages/signup_step2_page.dart';
+import 'package:student_apk/screens/sign_up_pages/signup_step3_page.dart';
+import 'package:student_apk/screens/sign_up_pages/signup_step4_page.dart';
+import 'package:student_apk/screens/sign_up_pages/verification_complete_page.dart';
 import 'package:student_apk/screens/dashboard_page.dart';
 import 'package:student_apk/screens/profile_page.dart';
 import 'package:student_apk/screens/settings_page.dart';
-import 'package:student_apk/screens/materials_page.dart';
-import 'package:student_apk/screens/subject_page.dart';
-import 'package:student_apk/screens/subject_content_page.dart';
-import 'package:student_apk/screens/files_page.dart';
-import 'package:student_apk/screens/forgot_password_step1.dart';
-import 'package:student_apk/screens/forgot_password_step2.dart';
-import 'package:student_apk/screens/forgot_password_step3.dart';
+import 'package:student_apk/screens/materials_pages/materials_page.dart';
+import 'package:student_apk/screens/materials_pages/subject_page.dart';
+import 'package:student_apk/screens/materials_pages/subject_content_page.dart';
+import 'package:student_apk/screens/materials_pages/files_page.dart';
+import 'package:student_apk/screens/forgot_pass_pages/forgot_password_step1.dart';
+import 'package:student_apk/screens/forgot_pass_pages/forgot_password_step2.dart';
+import 'package:student_apk/screens/forgot_pass_pages/forgot_password_step3.dart';
 import 'package:student_apk/theme.dart';
 
 void main() {
