+++ frontend/student_apk/lib/screens/forgot_pass_pages/forgot_password_step1.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import '../theme.dart';
-import '../routes.dart';
+import '../../theme.dart';
+import '../../routes.dart';
 
 class ForgotPasswordStep1 extends StatelessWidget {
   const ForgotPasswordStep1({super.key});
