+++ frontend/student_apk/lib/screens/forgot_pass_pages/forgot_password_step2.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import '../theme.dart';
-import '../routes.dart';
+import 'package:student_apk/theme.dart';
+import 'package:student_apk/routes.dart';
 
 class ForgotPasswordStep2 extends StatefulWidget {
   const ForgotPasswordStep2({super.key});
