+++ frontend/student_apk/lib/screens/materials_pages/materials_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import '../theme.dart';
-import '../routes.dart';
+import '../../theme.dart';
+import '../../routes.dart';
 
 class MaterialsPage extends StatelessWidget {
   const MaterialsPage({super.key});
