+++ frontend/student_apk/lib/screens/materials_pages/subject_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import '../theme.dart';
-import '../routes.dart';
+import '../../theme.dart';
+import '../../routes.dart';
 
 class SubjectPage extends StatelessWidget {
   const SubjectPage({super.key});
