+++ frontend/student_apk/lib/screens/sign_up_pages/verification_complete_page.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
-import '../theme.dart';
-import '../routes.dart';
+import 'package:student_apk/theme.dart';
+import 'package:student_apk/routes.dart';
 
 class VerificationCompletePage extends StatefulWidget {
   const VerificationCompletePage({super.key});
