+++ frontend/teacher_apk/lib/Screens/loading.dart
@@ -25,7 +25,7 @@ class _LoadingScreenState extends State<LoadingScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: Container( // Wrap body with Container for gradient
+      body: Container(
         decoration: const BoxDecoration(
           gradient: LinearGradient(
             colors: [AppTheme.backgroundGradientStart, AppTheme.backgroundGradientEnd],
@@ -37,23 +37,23 @@ class _LoadingScreenState extends State<LoadingScreen> {
           children: <Widget>[
             Align(
               alignment: Alignment.center,
-              child: const TriangleLogo(size: 100, isWhite: true), // Changed to white
+              child: TriangleLogo(size: 100, isWhite: true),
             ),
             Align(
-              alignment: Alignment.center, // This centers the Column horizontally
+              alignment: Alignment.center,
               child: Column(
-                mainAxisSize: MainAxisSize.min, // So it only takes needed vertical space
+                mainAxisSize: MainAxisSize.min,
                 children: <Widget>[
-                  const SizedBox(height: (80 / 2) + 380), // 40 + 380 = 420. Should be (100/2) + new_spacing
+                  const SizedBox(height: 430),
                   const ArcanumLogo(fontSize: 40, color: AppTheme.primaryColor),
                   const SizedBox(height: 10),
                   const Text(
                     "Teach smarter. Manage faster",
-                    style: TextStyle(fontSize: 14, fontFamily: 'Poppins', color: AppTheme.primaryColor), // Added text with small font
+                    style: TextStyle(fontSize: 14, fontFamily: 'Poppins', color: AppTheme.primaryColor),
                   ),
-                  const SizedBox(height: 50), // Spacing after text, consistent with previous layout
-                  const CircularProgressIndicator(
-                    valueColor: AlwaysStoppedAnimation<Color>(AppTheme.primaryColor), // Use primaryColor
+                  const SizedBox(height: 50),
+                  CircularProgressIndicator(
+                    valueColor: AlwaysStoppedAnimation<Color>(AppTheme.primaryColor),
                   ),
                 ],
               ),
